requirements.txt
# apache-airflow-providers-google>=14.0.0 
apache-airflow-providers-postgres>=6.1.3
apache-airflow-providers-common-sql>=1.26.0
pandas>=2.1.2